+++ linuxdoom-1.10/m_swap.c
@@ -26,19 +26,20 @@
 #endif
 #include "m_swap.h"
 
+#include <stdint.h>
 
 // Not needed with big endian.
 #ifndef __BIG_ENDIAN__
 
-// Swap 16bit, that is, MSB and LSB byte.
-unsigned short SwapSHORT(unsigned short x)
+// Swap 16bit, that is, MSB and LSB uint8_t.
+uint16_t SwapSHORT(uint16_t x)
 {
     // No masking with 0xFF should be necessary. 
     return (x>>8) | (x<<8);
 }
 
 // Swapping 32bit.
-unsigned long SwapLONG( unsigned long x)
+uint32_t SwapLONG( uint32_t x)
 {
     return
 	(x>>24)
